requirements.txt
streamlit==1.31.0
openai==1.12.0
openai-whisper==20231117
torch==2.2.0
torchaudio==2.2.0
numpy==1.24.3
ffmpeg-python==0.2.0